+++ src/com/haskforce/system/integrations/highlighting/HaskellAnnotationHolder.java
@@ -1,4 +1,4 @@
-package com.haskforce.haskell.highlighting.annotation;
+package com.haskforce.system.integrations.highlighting;
 
 import com.intellij.codeInsight.daemon.impl.AnnotationHolderImpl;
 import com.intellij.lang.annotation.Annotation;
