+++ src/com/haskforce/system/integrations/highlighting/HaskellProblem.java
@@ -1,6 +1,5 @@
 package com.haskforce.system.integrations.highlighting;
 
-import com.haskforce.haskell.highlighting.annotation.HaskellAnnotationHolder;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.PsiFile;
 import org.jetbrains.annotations.NotNull;
